+++ tests/index.test.ts
@@ -1,5 +1,7 @@
-import "dotenv/config";
-import PasteClient from "../src";
+import dotenv from "dotenv";
+import PasteClient from "../dist/index.js";
+
+dotenv.config();
 
 const KEY = process.env["TEST_API_KEY"];
 const USER_NAME = process.env["TEST_USER_NAME"];
@@ -18,18 +20,18 @@ const client = new PasteClient(KEY);
 async function test() {
   const token = await client.login(USER_NAME, USER_PASSWORD);
 
-  const deleted = await client.deletePasteByKey({
-    pasteKey: "XawZVJdA",
-    userKey: token,
-  });
-
-  console.log(deleted);
-
-  // const pastes = await client.getPastesByUser({
+  // const deleted = await client.deletePasteByKey({
+  //   pasteKey: "XawZVJdA",
   //   userKey: token,
   // });
 
-  // console.log(pastes);
+  // console.log(deleted);
+
+  const pastes = await client.getPastesByUser({
+    userKey: token,
+  });
+
+  console.log(pastes);
 
   // const url = await client.createPaste({
   //   code: "const x = 'hello world!'",